breaking change(react): remove default config for antd/arco
