+++ packages/plugin-react/src/react.ts
@@ -7,13 +7,12 @@ import {
 import type { RsbuildPluginAPI } from '@rsbuild/core';
 import type { PluginReactOptions } from '.';
 
-export const REACT_REFRESH_PATH = require.resolve('react-refresh');
-const REACT_REFRESH_DIR_PATH = path.dirname(REACT_REFRESH_PATH);
-
 export const applyBasicReactSupport = (
   api: RsbuildPluginAPI,
   options: PluginReactOptions,
 ) => {
+  const REACT_REFRESH_PATH = require.resolve('react-refresh');
+
   api.modifyBundlerChain(async (chain, { CHAIN_ID, isDev, isProd, target }) => {
     const config = api.getNormalizedConfig();
     const usingHMR = isUsingHMR(config, { isProd, target });
@@ -41,7 +40,7 @@ export const applyBasicReactSupport = (
       return;
     }
 
-    chain.resolve.alias.set('react-refresh', REACT_REFRESH_DIR_PATH);
+    chain.resolve.alias.set('react-refresh', path.dirname(REACT_REFRESH_PATH));
 
     const { default: ReactRefreshRspackPlugin } = await import(
       '@rspack/plugin-react-refresh'
